[project @ 2002-11-18 17:01:01 by simonpj] ----------------------------- 	Perl and control-M's on Win32 	-----------------------------
If the Cygwin 'perl' processes a file produced by a Mingw program
(such as ghc -M), so that the latter has DOS-style line termination,
the Cygwin 'perl' gets confused and adds extra control-M's.

This affects
	a) mkdependC, which processes the .depend file produced by
		ghc -M
	b) the perl post-procssing step in ghc/compiler/Makefile,
		which makes .depend-1 etc from .depend-BASE

The solution is to sprinkle few
	binmode( HANDLE );
statements around in the perl scripts.

I think they don't affect the Unix version.