[project @ 2002-11-21 09:36:03 by simonpj] ------------------------------- 	Fix and tidy the desugaring of 	pattern-matching in do-notation 	-------------------------------
In the reorgansiation of HsPat, failureFreePat had become incorrect (due to
a catch-all case that caught a constructor that should have been matched).  So
pattern-match failure in do-notation wasn't handled right.

As it turned out, DsExpr.dsDo could be made much simpler and more elegant
by using matchSimply instead of matchWrapper, and this had the side benefit
of removing the last call to HsPat.failureFreePat.  So it's gone!