[project @ 2003-02-04 15:31:18 by simonpj] ------------------------------------- 	Fix a name-capture bug in Ext-Core 	-------------------------------------
Don't expand newtypes (even non-recursive ones) when going to External Core.
Reason: the expansion was performed *after* Tidying; the expansion performs
type substitution, which is only done right if you take account of the Uniques.
But since it's post-tidying, we got capture of occurence names.

I hope the lack of newtype expansion doesn't hurt anyone; I doubt it will.
If so, we can think again.

Thanks to Tobias Gedell for this one.