[project @ 2003-02-12 15:01:31 by simonpj] -------------------------------------   Big upheaval to the way that constructors are named 	-------------------------------------
This commit enshrines the new story for constructor names.  We could never
really get External Core to work nicely before, but now it does.

The story is laid out in detail in the Commentary
	ghc/docs/comm/the-beast/data-types.html
so I will not repeat it here.

	[Manuel: the commentary isn't being updated, apparently.]

However, the net effect is that in Core and in External Core, contructors look
like constructors, and the way things are printed is all consistent.

It is a fairly pervasive change (which is why it has been so long postponed),
but I hope the question is now finally closed.

All the libraries compile etc, and I've run many tests, but doubtless there will
be some dark corners.