[project @ 2003-03-10 13:27:34 by simonmar] Use STK_CHK_NP rather than STK_CHK_GEN in the entry code for AP and AP_STACK thunks.
The reason is that if the check fails, on return the closure might
have been blackholed (perhaps due to lazy blackholing, or even because
another thread has entered it), so we *must* enter the closure to
continue, rather than just continuing at the original entry point.

All compiler-generated thunks do this anyway, I think these are the
only cases of thunks which don't follow the rule.

Thanks to Robert Ennals for spotting, and correctly diagnosing, the
bug.