[project @ 2003-03-27 08:21:27 by simonpj] -------------------------------------       Respect SPECIALISE pragmas in instance decls 	-------------------------------------
For some time now we have simply been discarding SPECIALISE pragmas
in instance declarations.  I think this was my fault, at some point
when I was re-plumbing TcClassDcl.lhs, but it's been this way for
some time.

The only uses of this facility in the Prelude are in GHC/Float.lhs
and GHC/Real.lhs, which affected the efficiency of the Float and
Double instance of properFraction, floor, ceiling, etc.

Ah well, it's fixed now.