[project @ 2003-03-31 14:02:32 by simonmar] On 32-bit platforms, add a bunch of symbols from libgcc.a used to support 64-bit arithmetic.
It looks like these symbols are defined in libc on some platforms but
not others, but it shouldn't do any harm to include them in the list
of symbols that the linker knows about even if they're also in libc.
They were already included for some platforms (cygwin, mingw, and
darwin), this commit just includes then whenever we're compiling with
gcc on a 32-bit platform.