[project @ 2003-04-08 13:06:48 by simonpj] ----------------------------------    Try swapping order of inference for applications 	----------------------------------
For a long time we've inferred the argument types of a function
call before using information from the expected type passed in.
This commit swaps the order.  See comments with tcApp for an
example where this makes a real difference.

I don't think any error messages get signficantly worse, but yell
if you find any.