[project @ 2003-04-09 08:07:58 by simonpj] -------------------------------------       Fix a functional-dependency-related bug in 		tcSimpifyRestricted 	-------------------------------------
	MERGE TO STABLE if it goes over conveniently
		(but I rather think it may not)

tcSimplifyRestricted works by (a) simplifying brutall to find out
what the constrained type variables are, and (b) simplifying more
gently, knowing the constrained type varaibles.  The bug is that
in step (b) we were not doing the check-for-improvement-and-loop
part, thinking that step (a) had alrady done all the improvement.
But not so, as an example in the code now shows.

Simple to fix.  I rather think we could tidy up these various loops.