[project @ 2003-04-10 14:44:18 by simonpj] ----------------------------------        Fix a long-standing eta-reduction bug 	----------------------------------
Consider the stupid definition

	f = \x -> f x

We were erroneously eta-reducing this to

	f = f

(unsound because they'd be distinguishable by `seq`)

The reason was that simplLazyBind was exposing the arity of
a recursive function to its own RHS, when all it was really
trying to do was expose the *rules* for the function.

Easily fixed.   This fixes some

	"Bad eta expand"

warnings.  Good all round.  In particular, fixes rn006.