[project @ 2003-05-21 13:05:49 by simonmar] Restore the correct partitioning of command-line arguments into objects and compilation-manager inputs for --make mode, and restore a comment explaining the behaviour.
Rev. 1.120 made some subtle changes to the semantics, in particular in
cases where a file given on the command line is neither a source nor
an object file.

I believe the behaviour for one-shot mode has not changed.  The
behaviour for GHCi mode is now the same as --make mode (previous to
rev. 1.120 it was subtly broken in this respect).