[project @ 2003-05-23 14:20:35 by simonmar] This test is an expected failure, but its original purpose seems to have been lost along the way somewhere.
It is indeed failing, but I susepct not for the same reasons that it
was originally failing.  After fixing a few compilation errors I
managed to get it to run, whereupon it fails with "Prelude.read: no
parse", which may or may not be the correct result.  As far as I can
tell, this test never worked.  In the previous incarnation of the test
suite, it was even disabled.